refactor: 改 `plugins/` 文件夹为 `features/`
- 将 `src/plugins` 目录重命名为 `src/features`
- 同步更新导入

认为插件没有插拔需求，可以叫做功能（Features）
