+++ src/features/autoFishing.ts
