+++ src/features/killedBySimPlayer.ts
